♻️ refactor : 적성 검사 이력 관련 로직을 별도 서비스로 분리
[변경 사항]
- AptitudeTestHistoryService 추가
  - 테스트 이력 저장/조회/삭제 로직 이관
  - 세션 관리 및 완료 처리 담당

- AptitudeService 리팩토링
  - 적성 검사 핵심 비즈니스 로직만 유지
  - 이력 관련 작업은 AptitudeTestHistoryService에 위임
  - Entity Builder 패턴으로 데이터 생성 후 전달

- AptitudeController 수정
  - TestHistoryService 주입 추가
  - 이력 조회 API는 HistoryService 직접 호출

[개선 효과]
- 단일 책임 원칙(SRP) 준수
- 서비스 간 책임 분리로 유지보수성 향상
- 코드 가독성 및 테스트 용이성 개선

Resolves: #164 #166
